move token generator
